fixed a numerical resolution bug that was causing it to find the wrong X-axis index for a given x-axis value (cast to float was rounding to wrong value)

+++ src/org/opensha/commons/data/function/EvenlyDiscretizedFunc.java
@@ -273,7 +273,7 @@ public class EvenlyDiscretizedFunc extends AbstractDiscretizedFunc{
 	 */
 	public int getXIndex( double x) throws Point2DException{
 
-		int i = Math.round((float)((x-minX)/delta));
+		int i = (int)Math.round((double)((x-minX)/delta));
 		if( withinTolerance(x, this.getX(i) ))
 			return i;
 		else
